Add BeaconSet update count calculations, extract parameters to file, re-combine scripts
