+++ scripts/dapi-pricing-parameters.json
@@ -1,0 +1,51 @@
+{
+  "startDate": "2023-10-30T07:59:05.000Z",
+  "endDate": "2023-12-30T09:59:05.000Z",
+  "updateCountOptions": [
+    {
+      "deviationThreshold": 0.25,
+      "heartbeatInterval": 120
+    },
+    {
+      "deviationThreshold": 0.25,
+      "heartbeatInterval": 86400
+    },
+    {
+      "deviationThreshold": 0.5,
+      "heartbeatInterval": 86400
+    },
+    {
+      "deviationThreshold": 1.0,
+      "heartbeatInterval": 86400
+    }
+  ],
+  "updateCountCheckFrequency": 10,
+  "defaultSingleUpdateGasCost": 173000,
+  "defaultGasMultiplier": 1.5,
+  "defaultSubscriptionDuration": 183,
+  "chainSubscriptionDurations": {},
+  "chainGasMultipliers": {},
+  "chainSingleUpdateGasCosts": {
+    "10": { "fixedGasCost": 130000, "l1FixedGasCost": 25000 },
+    "420": { "fixedGasCost": 130000, "l1FixedGasCost": 25000 },
+    "324": { "fixedGasCost": 5400000 },
+    "1101": { "fixedGasCost": 180000 },
+    "1442": { "fixedGasCost": 180000 },
+    "42161": { "fixedGasCost": 145000, "l1FixedGasCost": 15000 },
+    "421613": { "fixedGasCost": 145000, "l1FixedGasCost": 15000 },
+    "8453": { "fixedGasCost": 173000, "l1FixedGasCost": 27888 },
+    "84531": { "fixedGasCost": 173000, "l1FixedGasCost": 27888 },
+    "5000": { "fixedGasCost": 173000, "l1FixedGasCost": 1850 },
+    "5001": { "fixedGasCost": 173000, "l1FixedGasCost": 1850 }
+  },
+  "chainIdToGasOracleContractAddress": {
+    "534353": "0x5300000000000000000000000000000000000002",
+    "84531": "0x420000000000000000000000000000000000000F",
+    "8453": "0x420000000000000000000000000000000000000F",
+    "5001": "0x420000000000000000000000000000000000000F",
+    "5000": "0x420000000000000000000000000000000000000F",
+    "420": "0x420000000000000000000000000000000000000F",
+    "10": "0x420000000000000000000000000000000000000F"
+  },
+  "chainNativeTokenPrices": { "1": 2200, "5000": 0.62 }
+}
