--- src/database/database.js
@@ -1,13 +1,0 @@
-require('dotenv').config();
-const { Client } = require('pg');
-
-// Set the database credentials
-const client = new Client({
-  host: process.env.DATABASE_HOST,
-  port: process.env.DATABASE_PORT,
-  database: process.env.DATABASE_NAME,
-  user: process.env.DATABASE_USER,
-  password: process.env.DATABASE_PASSWORD,
-});
-
-module.exports = { client };
